fix(arith): C54.G-hybrid — trap on bare overflow, null-propagate inside `??`
Phase 1 of the C54 initiative lands the semantic fix for arithmetic
overflow and divide-by-zero.  Under G-hybrid:

- **Bare** arithmetic (`a + b`, `a / b`, `a % b` not wrapped in `??`)
  traps at runtime with `file:line:col` and the operator — no more
  silent-wrong-result via the `i32::MIN` null sentinel.
- **Inside a `?? default` wrapper** (only the immediate LHS of `??`),
  overflow/div-zero produces `i32::MIN` and `??` discharges to the
  RHS — the `x = (a * b) ?? fallback` idiom keeps working.
- Both paths are uniform across interpreter, native codegen, and
  WASM (WASM shares the interpreter; native dispatches through
  `ops::op_*` directly).

## Key changes

- `src/ops.rs`: `checked_int!` / `checked_long!` always trap (the
  release branch that returned the sentinel is gone).  New
  `checked_int_nullable!` / `checked_long_nullable!` macros produce
  the sentinel.  New `op_{add,sub,mul,div,rem}_int_nullable` +
  `_long_nullable` functions using the nullable macros.
- `default/01_code.loft`: 5 new `Op{Add,Min,Mul,Div,Rem}IntNullable`
  opcodes calling the new functions.  Long Nullable variants
  deferred pending Phase 5 (opcode reclamation).
- `src/parser/operators.rs::handle_operator` at `??`: when the LHS
  is a direct arithmetic Call, rewrite its def-number to the
  Nullable variant so overflow produces null instead of trapping.
  Only the outermost call is rewritten — nested sub-expressions
  still trap.
- `src/fill.rs`: regenerated via `regen_fill_rs`, OPERATORS array
  grown 255 → 260 to accommodate the new opcodes.
- `src/data.rs` + `src/state/debug.rs` + `src/compile.rs`: widen
  the opcode type from `u8` to `u16` for opcode-metadata lookup,
  HashMap keys, and disassembly.  The bytecode stream's 2-byte
  encoding (`[255, offset]` for opcodes ≥ 255) was already handled
  at the main `execute` loop, but `execute_log` in debug.rs and the
  disassembler in compile.rs only read a single byte.  Fixed both
  to read the escape sequence.

## Probe verification

- `probe_00 baseline`: `i32::MAX + 1` now **traps** (was silent-null).
- `probe_08 ?? on arithmetic`: `(i32::MAX * 2) ?? 42` returns **42**
  — G-hybrid preserves the idiom.
- `probes 01-06, 09`: now trap at the arithmetic overflow instead
  of silently flowing null into a `not null` slot.  H1-H4 holes
  remain for explicit null values but overflow can no longer reach.

## Test updates

- Four `ops::test::*_overflow` + `*_sentinel` tests un-gated from
  `#[cfg(debug_assertions)]` — traps now fire uniformly in release.
- `and_int_sentinel` stays debug-only; it tests `sentinel_int!`
  which is a separate bitwise path, not C54.G arithmetic.
- `p85b_*` + `p156_*` line-number expectations updated from
  `01_code.loft:385:24` to `403:24` (constants shifted by my 18
  added lines).

## Verification

- `cargo test --release --test leak` — 13 passed.
- `scripts/find_problems.sh --bg --wait` — full workspace 0
  failures.
- `cargo fmt -- --check` + `cargo clippy --release --all-targets --
  -D warnings` — clean.
- All 16 moros_sim suites + 4 moros_ui suites stay green.

## Follow-ups

- Phase 2 (C54.A) widens `integer` to i64 + bumps `.loftc`.
- Phase 5 (C54.E) reclaims ~28 `Op*Long` slots, freeing room to
  add the deferred Long Nullable variants.
- H1-H4 runtime `not null` enforcement is a separate multi-phase
  effort; a future C54.G′ migration depends on it.

+++ src/data.rs
@@ -805,7 +805,7 @@ pub struct Data {
     statics: Vec<u8>,
     pub(crate) op_codes: u16,
     possible: HashMap<String, Vec<u32>>,
-    pub(crate) operators: HashMap<u8, u32>,
+    pub(crate) operators: HashMap<u16, u32>,
     /// PKG.4: native function symbols — loft function name → Rust symbol path.
     /// Populated when packages with `[native.functions]` are loaded.
     /// Keys are the user-facing loft names (e.g. `save_png`), not the internal
@@ -1044,7 +1044,7 @@ impl Data {
             self.op_codes
         );
         self.definitions[def_nr as usize].op_code = self.op_codes;
-        self.operators.insert(self.op_codes as u8, def_nr);
+        self.operators.insert(self.op_codes, def_nr);
         self.op_codes += 1;
     }
 
@@ -1719,12 +1719,12 @@ impl Data {
     }
 
     #[must_use]
-    pub fn has_op(&self, op: u8) -> bool {
+    pub fn has_op(&self, op: u16) -> bool {
         self.operators.contains_key(&op)
     }
 
     #[must_use]
-    pub fn operator(&self, op: u8) -> &Definition {
+    pub fn operator(&self, op: u16) -> &Definition {
         self.def(self.operators[&op])
     }
